Move version to its own configuration package so it is reachable from outside main.
Add a cli-version key/value to ~/.fn/config.yaml in order to track the
version of the cli.
