+++ commands/version.go
@@ -1,0 +1,41 @@
+package commands
+
+import (
+	"fmt"
+
+	"github.com/fnproject/cli/client"
+	"github.com/fnproject/cli/config"
+	"github.com/urfave/cli"
+)
+
+// VersionCommand
+func VersionCommand() cli.Command {
+	return cli.Command{
+		Name:        "version",
+		Usage:       "Display CLI and server versions",
+		Description: "This is commands shows the latest client and server version.",
+		Action:      versionCMD,
+	}
+}
+
+func versionCMD(c *cli.Context) error {
+	provider, err := client.CurrentProvider()
+	if err != nil {
+		return err
+	}
+
+	ver := config.GetVersion("latest")
+	if ver == "" {
+		ver = "Client version: " + config.Version
+	}
+	fmt.Println(ver)
+
+	versionClient := provider.VersionClient()
+	v, err := versionClient.GetVersion(nil)
+	if err != nil {
+		fmt.Println("Server version: ", "?")
+		return nil
+	}
+	fmt.Println("Server version: ", v.Payload.Version)
+	return nil
+}
