+++ config/config.go
@@ -5,6 +5,7 @@ import (
 	"os"
 	"path/filepath"
 	"strings"
+
 	"github.com/fnproject/fn_go/provider"
 	"github.com/spf13/viper"
 	"github.com/urfave/cli"
@@ -22,18 +23,19 @@ const (
 
 	ReadWritePerms = os.FileMode(0755)
 
-	CurrentContext  = "current-context"
-	ContextProvider = "provider"
+	CurrentContext    = "current-context"
+	ContextProvider   = "provider"
+	CurrentCliVersion = "cli-version"
 
 	EnvFnRegistry = "registry"
 	EnvFnContext  = "context"
 )
 
-var defaultRootConfigContents = &ContextMap{CurrentContext: ""}
+var defaultRootConfigContents = &ContextMap{CurrentContext: "", CurrentCliVersion: ""}
 var DefaultContextConfigContents = &ContextMap{
-	ContextProvider: DefaultProvider,
-	provider.CfgFnAPIURL:     defaultLocalAPIURL,
-	EnvFnRegistry:   "",
+	ContextProvider:      DefaultProvider,
+	provider.CfgFnAPIURL: defaultLocalAPIURL,
+	EnvFnRegistry:        "",
 }
 
 type ContextMap map[string]string
@@ -128,9 +130,14 @@ func LoadConfiguration(c *cli.Context) error {
 	viper.AddConfigPath(filepath.Join(home, rootConfigPathName, contextsPathName))
 	viper.SetConfigName(context)
 
+	if viper.GetString("cli-version") != Version {
+		fmt.Printf("Welcome to Fn CLI %s - please check the release notes at %s \n\n", Version, "https://github.com/fnproject/cli/blob/master/README.md")
+		WriteCurrentContextToConfigFile(CurrentCliVersion, Version)
+	}
+
 	if err := viper.ReadInConfig(); err != nil {
 		fmt.Printf("%v \n", err)
-		err := WriteCurrentContextToConfigFile("default")
+		err := WriteCurrentContextToConfigFile(CurrentContext, "default")
 		if err != nil {
 			return err
 		}
@@ -142,7 +149,7 @@ func LoadConfiguration(c *cli.Context) error {
 	return nil
 }
 
-func WriteCurrentContextToConfigFile(value string) error {
+func WriteCurrentContextToConfigFile(key, value string) error {
 	home := GetHomeDir()
 
 	configFilePath := filepath.Join(home, rootConfigPathName, contextConfigFileName)
@@ -159,7 +166,7 @@ func WriteCurrentContextToConfigFile(value string) error {
 
 	configValues := ContextMap{}
 	for k, v := range *file {
-		if k == CurrentContext {
+		if k == key {
 			configValues[k] = value
 		} else {
 			configValues[k] = v
